concatenator: Replace with simple function
